fix(prefabs): allow option group buttons to be touched on and off
There was an issue where Option buttons would not toggle each other
on and off because the Select action was going through the
SpatialTarget and an Option group needs the Spatial Dispatcher to
dispatch the Select action so it knows which other buttons to dispatch
the Deselect action to.

The SpatialButton.OptionButton prefab now uses the
SpatialButtonConfigurator Select action as this goes through the
correct dispatcher when dealing with button groups.