fix(prefabs): force apply TextMeshPro styles when state is enabled
There is an issue in later versions of Unity where if the TextMeshPro
component is not enabled in the scene then it does not apply the
state styles to the TextMeshPro component. This fix uses a
BehaviourEnabledObserver to check to see whtn the TextMeshPro for
each style is enabled and then force applies the styles.