feat(Editor): add custom buttons for facade
A new custom inspector editor has been added for the Spatial Button
Facade that renders a `Copy Text and Size` button that will copy
the text and font size from the enabled inactive style to all the other
states to make it easier to update.

There is also a `Preview Style` button that will apply the styles in
the editor so it is easy to see what the button looks like without
needing to run the scene in play mode.