fix(Materials): swap button shader to mobile friendly version
The standard unity particle shaders do not behave well on mobile
devices such as the oculus quest so the shader has been switched out
on the SpatialButton material to the Tilia vertex unlit shader.

The style setting code has been updated now to just use a simple shared
material color set as this seems to work for all shaders.